requirements.txt
asyncpg>=0.29
jinja2>=3.1.0
mcp>=0.1.0
psutil>=7.1
rich>=13.7
pytest>=7.4
pytest-asyncio>=0.23
portalocker>=2.0
pyyaml>=6.0
watchdog>=3.0.0
